一点点宿舍模块

+++ src/main/java/com/we_are_team/school_background_service_system/service/Impl/LostAndFoundServiceImpl.java
@@ -4,11 +4,13 @@ import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.we_are_team.school_background_service_system.context.BaseContext;
 import com.we_are_team.school_background_service_system.mapper.LostAndFoundMapper;
+import com.we_are_team.school_background_service_system.mapper.UserMapper;
 import com.we_are_team.school_background_service_system.pojo.dto.AddCategoryDTO;
 import com.we_are_team.school_background_service_system.pojo.dto.AddLocationDTO;
 import com.we_are_team.school_background_service_system.pojo.entity.ItemCategory;
 import com.we_are_team.school_background_service_system.pojo.entity.LocationCategory;
 import com.we_are_team.school_background_service_system.pojo.entity.LostAndFound;
+import com.we_are_team.school_background_service_system.pojo.entity.User;
 import com.we_are_team.school_background_service_system.pojo.vo.GetLostAndFoundVO;
 import com.we_are_team.school_background_service_system.pojo.vo.ItemCategoryVO;
 import com.we_are_team.school_background_service_system.pojo.vo.LocationCategoryVO;
@@ -33,11 +35,20 @@ import java.util.UUID;
 @Service
 public class LostAndFoundServiceImpl implements LostAndFoundService {
     @Autowired
+    private UserMapper userMapper;
+    @Autowired
     private LostAndFoundMapper lostAndFoundMapper;
     @Autowired
     private AliOssUtil aliOssUtil;
+
+
     @Override
     public void addCategory(AddCategoryDTO addCategoryDTO) {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限添加物品分类");
+        }
+
         ItemCategory itemCategory = new ItemCategory();
         itemCategory.setCategoryName(addCategoryDTO.getCategoryName());
         itemCategory.setCreatedTime(LocalDateTime.now());
@@ -46,6 +57,10 @@ public class LostAndFoundServiceImpl implements LostAndFoundService {
 
     @Override
     public void addLocation(AddLocationDTO addLocationDTO) {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限添加地点分类");
+        }
         LocationCategory locationCategory = new LocationCategory();
         locationCategory.setLocationName(addLocationDTO.getLocationName());
         locationCategory.setCreatedTime(LocalDateTime.now());
@@ -54,22 +69,38 @@ public class LostAndFoundServiceImpl implements LostAndFoundService {
 
     @Override
     public void deleteCategory(Integer categoryId) {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限删除物品分类");
+        }
         lostAndFoundMapper.deleteCategory(categoryId);
     }
 
     @Override
     public void deleteLocation(Integer locationId) {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限删除地点分类");
+        }
         lostAndFoundMapper.deleteLocation(locationId);
     }
 
     @Override
     public List<ItemCategoryVO> getAllItemCategory() {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5") || !user.getPermission().contains("0")){
+            throw new RuntimeException("没有权限查看物品分类");
+        }
         List<ItemCategoryVO> itemCategoryVOList = lostAndFoundMapper.getAllItemCategory();
         return itemCategoryVOList;
     }
 
     @Override
     public List<LocationCategoryVO> getAllLocationCategory() {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5") || !user.getPermission().contains("0")){
+            throw new RuntimeException("没有权限查看地点分类");
+        }
         List<LocationCategoryVO> locationCategoryVOList = lostAndFoundMapper.getAllLocationCategory();
         return locationCategoryVOList;
     }
@@ -78,6 +109,16 @@ public class LostAndFoundServiceImpl implements LostAndFoundService {
  */
     @Override
     public void uploadLostFound(String itemName, Integer categoryId, Integer locationId, String description, MultipartFile[] imageUrlsArray) {
+        if(itemName == null || itemName.equals("")
+           || categoryId == null || categoryId == 0
+           || locationId == null || locationId == 0
+              ){
+            throw new RuntimeException("请填写完整信息");
+        }
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限上传失物招领信息");
+        }
         LostAndFound lostAndFound = new LostAndFound();
         lostAndFound.setItemName(itemName);
         lostAndFound.setAdminId(BaseContext.getCurrentId());
@@ -117,6 +158,16 @@ public class LostAndFoundServiceImpl implements LostAndFoundService {
 
     @Override
     public void updateLostFound(Integer itemId, String itemName, Integer categoryId, Integer locationId, String description, MultipartFile[] imageUrlsArray,Integer status) {
+        if(itemName == null || itemName.isEmpty()
+                || categoryId == null || categoryId == 0
+                || locationId == null || locationId == 0
+        ){
+            throw new RuntimeException("请填写完整信息");
+        }
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限修改失物招领信息");
+        }
         LostAndFound lostAndFound = new LostAndFound();
         lostAndFound.setItemId(itemId);
         lostAndFound.setItemName(itemName);
@@ -155,11 +206,19 @@ public class LostAndFoundServiceImpl implements LostAndFoundService {
 
     @Override
     public void deleteLostFound(Integer itemId) {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if(!user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限删除失物招领信息");
+        }
          lostAndFoundMapper.deleteLostFound(itemId);
     }
 
     @Override
     public PageResult getLostAndFoundList(Integer pageNum, Integer pageSize, String orderKey, LocalDate beginTime, LocalDate endTime, String status) {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if (!user.getPermission().contains("0") || !user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限查看失物招领信息");
+        }
         PageHelper pageHelper = new PageHelper();
         pageHelper.startPage(pageNum, pageSize);
         Page<GetLostAndFoundVO> lostAndFoundVOList = lostAndFoundMapper.getLostAndFoundList(orderKey,beginTime,endTime,status);
@@ -174,6 +233,10 @@ public class LostAndFoundServiceImpl implements LostAndFoundService {
 
     @Override
     public LostAndFoundVO getLostAndFoundDetail(Integer itemId) {
+        User user =  userMapper.getUserByUserId(BaseContext.getCurrentId());
+        if (!user.getPermission().contains("0") || !user.getPermission().contains("5")){
+            throw new RuntimeException("没有权限查看失物招领信息");
+        }
         LostAndFound lostAndFound = lostAndFoundMapper.getLostAndFoundDetail(itemId);
         LostAndFoundVO lostAndFoundVO = new LostAndFoundVO();
         BeanUtils.copyProperties(lostAndFound,lostAndFoundVO);
